+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.11.1] - 2025-10-17
+
+- Added
+  - Commands for generating SSA assertions and access tokens
+- Updated
+  - Users can now select OAuth scopes by checking checkboxes
+
 ## [2.11.0] - 2025-10-07
 
 - Added